refac: apply code review
