+++ src/main/java/auth/controller/dto/TokenRequest.java
@@ -1,4 +1,4 @@
-package auth.domain.dto;
+package auth.controller.dto;
 
 public class TokenRequest {
     private String username;
