+++ src/main/java/auth/repository/UserDetailsDao.java
@@ -1,6 +1,6 @@
 package auth.repository;
 
-import auth.domain.persist.UserDetails;
+import auth.domain.UserDetails;
 import lombok.RequiredArgsConstructor;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 @RequiredArgsConstructor
-public class UserDetailsRepositoryImpl implements UserDetailsRepository {
+public class UserDetailsDao {
     private final JdbcTemplate jdbcTemplate;
 
     private final RowMapper<UserDetails> rowMapper = (resultSet, rowNum) -> new UserDetails(
@@ -20,13 +20,11 @@ public class UserDetailsRepositoryImpl implements UserDetailsRepository {
             resultSet.getString("role")
     );
 
-    @Override
-    public UserDetails findById(Long id) {
+    public UserDetails findById(long id) {
         String sql = "SELECT id, username, password, name, phone, role from member where id = ?;";
         return jdbcTemplate.queryForObject(sql, rowMapper, id);
     }
 
-    @Override
     public UserDetails findByUsername(String username) {
         String sql = "SELECT id, username, password, name, phone, role from member where username = ?;";
         return jdbcTemplate.queryForObject(sql, rowMapper, username);
