--- src/main/java/auth/repository/UserDetailsRepository.java
@@ -1,9 +1,0 @@
-package auth.repository;
-
-import auth.domain.persist.UserDetails;
-
-public interface UserDetailsRepository {
-    UserDetails findById(Long id);
-
-    UserDetails findByUsername(String username);
-}
