+++ src/main/java/auth/service/LoginService.java
@@ -1,9 +1,9 @@
 package auth.service;
 
-import auth.domain.dto.TokenRequest;
-import auth.domain.dto.TokenResponse;
-import auth.domain.persist.UserDetails;
-import auth.repository.UserDetailsRepository;
+import auth.controller.dto.TokenRequest;
+import auth.controller.dto.TokenResponse;
+import auth.domain.UserDetails;
+import auth.repository.UserDetailsDao;
 import auth.support.AuthenticationException;
 import auth.support.JwtTokenProvider;
 import lombok.RequiredArgsConstructor;
@@ -12,12 +12,12 @@ import org.springframework.stereotype.Service;
 @Service
 @RequiredArgsConstructor
 public class LoginService {
-    private final UserDetailsRepository userDetailsRepository;
+    private final UserDetailsDao userDetailsDao;
     private final JwtTokenProvider jwtTokenProvider;
 
     public TokenResponse createToken(TokenRequest tokenRequest) {
-        UserDetails userDetails = userDetailsRepository.findByUsername(tokenRequest.getUsername());
-        if (userDetails == null || userDetails.checkWrongPassword(tokenRequest.getPassword())) {
+        UserDetails userDetails = userDetailsDao.findByUsername(tokenRequest.getUsername());
+        if (userDetails == null || !userDetails.isValidPassword(tokenRequest.getPassword())) {
             throw new AuthenticationException();
         }
 
@@ -26,12 +26,12 @@ public class LoginService {
         return new TokenResponse(accessToken);
     }
 
-    public Long extractPrincipal(String credential) {
+    public long extractPrincipal(String credential) {
         return Long.parseLong(jwtTokenProvider.getPrincipal(credential));
     }
 
-    public UserDetails extractMember(String credential) {
-        Long id = Long.parseLong(jwtTokenProvider.getPrincipal(credential));
-        return userDetailsRepository.findById(id);
+    public UserDetails extractUserDetails(String credential) {
+        long id = Long.parseLong(jwtTokenProvider.getPrincipal(credential));
+        return userDetailsDao.findById(id);
     }
 }
