+++ src/main/java/auth/support/LoginMemberArgumentResolver.java
@@ -1,6 +1,6 @@
 package auth.support;
 
-import auth.domain.template.LoginMember;
+import auth.support.template.LoginMember;
 import auth.service.LoginService;
 import org.springframework.core.MethodParameter;
 import org.springframework.http.HttpHeaders;
@@ -25,7 +25,7 @@ public class LoginMemberArgumentResolver implements HandlerMethodArgumentResolve
     public Object resolveArgument(MethodParameter parameter, ModelAndViewContainer mavContainer, NativeWebRequest webRequest, WebDataBinderFactory binderFactory) throws Exception {
         try {
             String credential = webRequest.getHeader(HttpHeaders.AUTHORIZATION).split(" ")[1];
-            return loginService.extractMember(credential);
+            return loginService.extractUserDetails(credential);
         } catch (Exception e) {
             throw new AuthenticationException();
         }
