+++ src/main/java/nextstep/domain/Schedule.java
@@ -1,4 +1,4 @@
-package nextstep.domain.persist;
+package nextstep.domain;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
