+++ src/main/java/nextstep/domain/Waiting.java
@@ -1,36 +1,24 @@
-package nextstep.domain.persist;
+package nextstep.domain;
+
+import lombok.Getter;
 
 import java.util.Objects;
 
-public class Reservation {
+@Getter
+public class Waiting {
     private Long id;
     private Schedule schedule;
     private Member member;
 
-    public Reservation() {
-    }
-
-    public Reservation(Schedule schedule, Member member) {
+    public Waiting(Schedule schedule, Member member) {
         this.schedule = schedule;
         this.member = member;
     }
 
-    public Reservation(Long id, Schedule schedule, Member member) {
+    public Waiting(Long id, Schedule schedule, Member member) {
         this.id = id;
         this.schedule = schedule;
         this.member = member;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public Schedule getSchedule() {
-        return schedule;
-    }
-
-    public Member getMember() {
-        return member;
     }
 
     public boolean sameMember(Member member) {
