+++ src/main/java/nextstep/repository/ReservationDao.java
@@ -1,9 +1,9 @@
 package nextstep.repository;
 
-import nextstep.domain.persist.Member;
-import nextstep.domain.persist.Reservation;
-import nextstep.domain.persist.Schedule;
-import nextstep.domain.persist.Theme;
+import nextstep.domain.Member;
+import nextstep.domain.Reservation;
+import nextstep.domain.Schedule;
+import nextstep.domain.Theme;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
