+++ src/main/java/nextstep/repository/ThemeDao.java
@@ -1,6 +1,6 @@
 package nextstep.repository;
 
-import nextstep.domain.persist.Theme;
+import nextstep.domain.Theme;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
