+++ src/main/java/nextstep/repository/WaitingDao.java
@@ -1,10 +1,10 @@
 package nextstep.repository;
 
 import lombok.RequiredArgsConstructor;
-import nextstep.domain.persist.Member;
-import nextstep.domain.persist.Schedule;
-import nextstep.domain.persist.Theme;
-import nextstep.domain.persist.Waiting;
+import nextstep.domain.Member;
+import nextstep.domain.Schedule;
+import nextstep.domain.Theme;
+import nextstep.domain.Waiting;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
