+++ src/test/java/nextstep/auth/AuthE2ETest.java
@@ -1,10 +1,10 @@
 package nextstep.auth;
 
-import auth.domain.dto.TokenRequest;
-import auth.domain.dto.TokenResponse;
+import auth.controller.dto.TokenRequest;
+import auth.controller.dto.TokenResponse;
 import io.restassured.RestAssured;
-import nextstep.domain.dto.MemberRequest;
-import nextstep.domain.dto.ThemeRequest;
+import nextstep.controller.dto.MemberRequest;
+import nextstep.controller.dto.ThemeRequest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
