+++ src/test/java/nextstep/reservation/ReservationE2ETest.java
@@ -4,11 +4,11 @@ import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import nextstep.AbstractE2ETest;
-import nextstep.domain.dto.ReservationRequest;
-import nextstep.domain.dto.ReservationResponse;
-import nextstep.domain.dto.ScheduleRequest;
-import nextstep.domain.dto.ThemeRequest;
-import nextstep.domain.persist.Reservation;
+import nextstep.controller.dto.ReservationRequest;
+import nextstep.controller.dto.ReservationResponse;
+import nextstep.controller.dto.ScheduleRequest;
+import nextstep.controller.dto.ThemeRequest;
+import nextstep.domain.Reservation;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
