+++ src/test/java/nextstep/theme/ThemeDaoTest.java
@@ -1,6 +1,6 @@
 package nextstep.theme;
 
-import nextstep.domain.persist.Theme;
+import nextstep.domain.Theme;
 import nextstep.repository.ThemeDao;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
