+++ src/test/java/nextstep/theme/ThemeE2ETest.java
@@ -1,11 +1,11 @@
 package nextstep.theme;
 
-import auth.domain.dto.TokenRequest;
-import auth.domain.dto.TokenResponse;
+import auth.controller.dto.TokenRequest;
+import auth.controller.dto.TokenResponse;
 import io.restassured.RestAssured;
 import nextstep.AbstractE2ETest;
-import nextstep.domain.dto.MemberRequest;
-import nextstep.domain.dto.ThemeRequest;
+import nextstep.controller.dto.MemberRequest;
+import nextstep.controller.dto.ThemeRequest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
